Fix shorter pages having different width.
This was due to the scrollbar appearing on
longer pages & making them narrower. The solution
was to make the scrollbar always-on.

Also fixed the post list appearing differently in
Chrome, due to it wrapping the date. I guess I will
have to test everything in both browsers from now on.